Fix #628 #630 reconction of Extension
Create a new internal service for extension
Create a tool for target reconnection
Create a tool for source reconnection
Use same rules for reconnection as creation tool

As tool are implemented for Extension such as #589, the reconnection of
association andGeneralization are available.
